requirements.txt
aiodns==3.0.0
aiocache==0.11.1
aiogram==2.17.1
aiohttp==3.8.1
aiosqlite==0.17.0
asyncpg==0.25.0
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
frozenlist==1.2.0
idna==3.3
iso8601==0.1.16
multidict==5.2.0
pycares==4.1.2
pycparser==2.21
pypika-tortoise==0.1.1
pytz>=2021.3
tortoise-orm==0.17.8
ujson==4.3.0
uvloop==0.16.0
